+++ dev/helm/release.sh
@@ -1,3 +1,3 @@
 cd ./docs/helm
-helm package ../../dev/helm/lucit-ubdcc/
-helm repo index . --url https://unicorn-depthcache-cluster-for-binance.docs.lucit.tech/helm
+helm package ../../dev/helm/ubdcc/
+helm repo index . --url https://oliver-zehentleitner.github.io/unicorn-depthcache-cluster-for-binance/helm