+++ packages/ubdcc-dcn/start_ubdcc_dcn.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+# ¯\_(ツ)_/¯
+#
+# File: packages/ubdcc-dcn/start_ubdcc_dcn.py
+#
+# Project website: https://github.com/oliver-zehentleitner/unicorn-depthcache-cluster-for-binance
+# Github: https://github.com/oliver-zehentleitner/unicorn-depthcache-cluster-for-binance
+# Documentation: https://oliver-zehentleitner.github.io/unicorn-depthcache-cluster-for-binance
+# PyPI: https://pypi.org/project/unicorn-depthcache-cluster-for-binance
+#
+# License: MIT
+# https://github.com/oliver-zehentleitner/unicorn-depthcache-cluster-for-binance/blob/master/LICENSE
+#
+# Author: Oliver Zehentleitner
+#
+# Copyright (c) 2024-2026, Oliver Zehentleitner (https://about.me/oliver-zehentleitner)
+# All rights reserved.
+
+import os
+from ubdcc_dcn import DepthCacheNode
+
+DepthCacheNode.DepthCacheNode(cwd=os.path.dirname(os.path.abspath(__file__)))