+++ dev/sphinx/source/changelog.md
@@ -5,15 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-[Discussions about unicorn-depthcache-cluster-for-binance releases!](https://github.com/LUCIT-Systems-and-Development/unicorn-depthcache-cluster-for-binance/discussions/categories/releases)
+[Discussions about unicorn-depthcache-cluster-for-binance releases!](https://github.com/oliver-zehentleitner/unicorn-depthcache-cluster-for-binance/discussions/categories/releases)
 
-[How to upgrade to the latest version!](https://unicorn-depthcache-cluster-for-binance.docs.lucit.tech/readme.html#installation-and-upgrade)
+[How to upgrade to the latest version!](https://oliver-zehentleitner.github.io/unicorn-depthcache-cluster-for-binance/readme.html#installation-and-upgrade)
 
 ## 0.1.4.dev (development stage/unreleased/unstable)
 
 ## 0.1.4
 ### Fixed
-- https://github.com/LUCIT-Systems-and-Development/unicorn-depthcache-cluster-for-binance/issues/8 - We don't overwrite 
+- https://github.com/oliver-zehentleitner/unicorn-depthcache-cluster-for-binance/issues/8 - We don't overwrite 
   the None anymore.
 
 ## 0.1.3